pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tokenwatch"
version = "0.1.0"
description = "Real-time API token usage proxy for Anthropic and OpenAI-compatible endpoints"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "rich>=13.0",
    "click>=8.1",
    "python-dotenv>=1.0",
]

[project.scripts]
tokenwatch = "tokenwatch.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
